+++ srcs/Utils/Types_parser.cpp
@@ -20,7 +20,7 @@ bool	parse_types_file(HashTable *file_formatname, const char *src)
 			continue;
 		if (formatname.size() && (formatname.size() != line.size()))
 		{
-			for (size_t i = 0; i < (std::count(line.begin(), line.end(), ' ') + 1); i++)
+			for (int i = 0; i < (std::count(line.begin(), line.end(), ' ') + 1); i++)
 			{
 				std::string occurrence;
 				currentline >> occurrence;
